+++ config/initializers/gitlab.rb
@@ -1,6 +1,6 @@
 # frozen_string_literal: true
 
 Gitlab.configure do |config|
-  config.endpoint       = ENV['GITLAB_ENDPOINT'] || 'https://gitlab.com/api/v4'
+  config.endpoint       = "https://#{ENV['GITLAB_HOST'] || 'gitlab.com'}/api/v4"
   config.private_token  = ENV['GITLAB_API_PRIVATE_TOKEN']
 end